requirements.txt
beautifulsoup4
requests
Mastodon.py
